Built non-flexible version of webpage.
Built initial version of webpage. There are the required text and image elements present but the current
version of the webpage does not flex and shrink as required. For example, the hero image will overflow
if the container is too small and the pictures in the body do not wrap.

The next iteration will be to fix these issues and to iron out any visual imperfections.